fix: some issue with return statement
